requirements.txt
feedparser>=6.0.0
httpx>=0.27.0
pyyaml>=6.0
python-dateutil>=2.9.0
anthropic>=0.40.0
